refactor(account): pull account methods into own file
